requirements.txt
ag-ui-protocol==0.1.9
aiohappyeyeballs==2.6.1  
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0   
anthropic==0.68.0        
anyio==4.10.0
argcomplete==3.6.2       
attrs==25.3.0
boto3==1.40.35
botocore==1.40.35        
cachetools==5.5.2        
certifi==2025.8.3        
charset-normalizer==3.4.3
click==8.3.0
cohere==5.18.0
colorama==0.4.6
coloredlogs==15.0.1      
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0 
email-validator==2.3.0   
eval_type_backport==0.2.2
executing==2.2.1
fastapi==0.117.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.0
fastavro==1.12.0
filelock==3.19.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.9.0
genai-prices==0.0.27
google-auth==2.40.3
google-genai==1.38.0
googleapis-common-protos==1.70.0
griffe==1.14.0
groq==0.31.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.35.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
invoke==2.2.0
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
logfire==4.8.0
logfire-api==4.8.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mcp==1.14.1
mdurl==0.1.2
mistralai==1.9.10
mpmath==1.3.0
multidict==6.6.4
nexus-rpc==1.1.0
numpy==2.2.6
onnxruntime==1.22.1
openai==1.108.1
opencv-python==4.12.0.88
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-httpx==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
openwakeword==0.6.0
packaging==25.0
pi==0.1.2
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pydantic==2.11.9
pydantic-ai==1.0.10
pydantic-ai-slim==1.0.10
pydantic-evals==1.0.10
pydantic-graph==1.0.10
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyperclip==1.10.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==311
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.27.1
rpi_hardware_pwm==0.3.0
rsa==4.9.1
s3transfer==0.14.0
scikit-learn==1.7.2
scipy==1.16.2
sentry-sdk==2.38.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.48.0
sympy==1.14.0
temporalio==1.17.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
tqdm==4.67.1
typer==0.19.1
types-protobuf==6.32.1.20250918
types-requests==2.32.4.20250913
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.36.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0